scripts/requirements.txt
yfinance
supabase
numpy
scipy
python-dotenv